some best practice

+++ index.js
@@ -112,8 +112,8 @@ async function run() {
         total_amount: orderedService.price,
         currency: query.currency,
         tran_id: transactionId, // use unique tran_id for each api call
-        success_url: `http://localhost:5000/payment/success?transactionId=${transactionId}`,
-        fail_url: "http://localhost:5000/payment/fail",
+        success_url: `${process.env.SERVER_URL}/payment/success?transactionId=${transactionId}`,
+        fail_url: `${process.env.SERVER_URL}/payment/fail?transactionId=${transactionId}`,
         cancel_url: "http://localhost:5000/payment/cancel",
         ipn_url: "http://localhost:3030/ipn",
         shipping_method: "Courier",
@@ -158,6 +158,9 @@ async function run() {
 
     app.post("/payment/success", async (req, res) => {
       const { transactionId } = req.query;
+      if (!transactionId) {
+        return res.redirect(`${process.env.CLIENT_URL}/payment/fail`);
+      }
       const result = await orderCollection.updateOne(
         { transactionId },
         {
@@ -169,16 +172,19 @@ async function run() {
       );
       if (result.modifiedCount > 0) {
         res.redirect(
-          `http://localhost:3000/payment/success?transactionId=${transactionId}`
+          `${process.env.CLIENT_URL}/payment/success?transactionId=${transactionId}`
         );
       }
     });
 
     app.post("/payment/fail", async (req, res) => {
       const { transactionId } = req.query;
+      if (!transactionId) {
+        return res.redirect(`${process.env.CLIENT_URL}/payment/fail`);
+      }
       const result = await orderCollection.deleteOne({ transactionId });
       if (result.deletedCount) {
-        res.redirect("http://localhost:3000/payment/fail");
+        res.redirect(`${process.env.CLIENT_URL}/payment/fail`);
       }
     });
